second checked 14 work

+++ controllers/cards.js
@@ -25,10 +25,13 @@ module.exports.createCard = (req, res, next) => {
 }
 
 module.exports.deleteCardById = (req, res, next) => {
+  if (req.params.cardId.length < 24) {
+    throw new ValidationError('Неправильный формат cardId')
+  }
   Card.findById(req.params.cardId)
     .then((card) => {
       if (card.owner != req.user._id) {
-        return Promise.reject(new Error('Вы можете удалить только свою карточку'));
+        throw new ForbiddenError('Вы можете удалить только свою карточку');
       }
       Card.findByIdAndRemove(req.params.cardId)
         .then(card => res.send({data: card}))
@@ -42,41 +45,73 @@ module.exports.deleteCardById = (req, res, next) => {
         });
     })
     .catch((err) => {
-      const error = new ForbiddenError('Вы можете только свою карточку');
-      next(err);
+      if (err.name === 'TypeError') {
+        const error = new CastError('Карточка с указанным _id не найдена')
+        next(error)
+      } else {
+        next(err);
+      }
     })
 }
 
 module.exports.likeCard = (req, res, next) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
-    { new: true },)
-  .then(card => res.send({data: card}))
-  .catch((err) => {
-    if (err.name === 'ValidationError') {
-      const error = new ValidationError("Переданны некорректные данные карточки")
-      next(error);
-    } else if (err.name === 'CastError') {
-      const error = new CastError('Карточка с указанным _id не найдена');
-      next(error);
-    } else {
-      next(err);
-    }
-  });
+  if (req.params.cardId.length < 24) {
+    throw new ValidationError('Неправильный формат cardId')
+  }
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+        { new: true },)
+      .then(card => res.send({data: card}))
+      .catch((err) => {
+        if (err.name === 'ValidationError') {
+          const error = new ValidationError("Переданны некорректные данные карточки")
+          next(error);
+        } else if (err.name === 'CastError') {
+          const error = new CastError('Карточка с указанным _id не найдена');
+          next(error);
+        } else {
+          next(err);
+        }
+      });
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        const error = new CastError('Карточка с указанным _id не найдена')
+        next(error)
+      } else {
+        next(err);
+      }
+    })
 }
 
 module.exports.dislikeCard = (req, res, next) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, // убрать _id из массива
-    { new: true },)
-  .then(card => res.send({data: card}))
-  .catch((err) => {
-    if (err.name === 'ValidationError') {
-      const error = new ValidationError("Переданны некорректные данные карточки")
-      next(error);
-    } else if (err.name === 'CastError') {
-      const error = new CastError('Карточка с указанным _id не найдена');
-      next(error);
-    } else {
-      next(err);
-    }
-  });
+  if (req.params.cardId.length < 24) {
+    throw new ValidationError('Неправильный формат cardId')
+  }
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, // убрать _id из массива
+        { new: true },)
+      .then(card => res.send({data: card}))
+      .catch((err) => {
+        if (err.name === 'ValidationError') {
+          const error = new ValidationError("Переданны некорректные данные карточки")
+          next(error);
+        } else if (err.name === 'CastError') {
+          const error = new CastError('Карточка с указанным _id не найдена');
+          next(error);
+        } else {
+          next(err);
+        }
+      });
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        const error = new CastError('Карточка с указанным _id не найдена')
+        next(error)
+      } else {
+        next(err);
+      }
+    })
 }
